requirements.txt
black==24.8.0
click==8.1.8
exceptiongroup==1.3.1
iniconfig==2.1.0
isort==5.13.2
mypy_extensions==1.1.0
packaging==26.0
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pytest==8.3.5
python-dotenv==1.0.1
tomli==2.4.0
typing_extensions==4.13.2
